fix protein construct names
In the last pull request, the strain names in the protein constructs
had `/` characters replaced with `_` which made them no longer match
the repo. This pull request fixes that by making a small
tweak to the `design_H3_HAconstructs.ipynb` notebook and then running
it to write FASTAs in `non-pipeline_analyses/library_design` with
the correct strain names.